+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.8.3 (unreleased)
 
+- Added support for `google-apis-monitoring_v3`
 - Added experimental support for Propshaft
 - Fixed error with walsender queries on live queries page